moved gem bundle to ${DATA_DIR}/tmp/ at migration
For user plugin installation it is absolutely necessary that
the gems are located in the data volume because some plugins
require gems to be installed.

Moving the gems bundle to the data volume will make sure that the gems
installed by plugins are available when the container is restarted.

Whenever there is a version update, the bundle directory will be removed
and the plugins gems would need to be reinstalled.