feature: create backups
Starting the container with `app:backup:create` or executing the script
`redmine-backup-create` creates a timestamped backup of the Redmine
installation in the `backups/` directory located at the persistent data
volume.

The generated backup consists of:
 - database
 - files (attachments)
 - plugins
 - themes
 - dotfiles
 - backup_information.yml

The `backup_information.yml` contains information about the Redmine
version and date when the backup was created.

Closes #125
Refer #110
Refer #132