feature: backup restore
This feature allows users to restore backups created by the
`app:backup:create` argument or the `redmine-backup-create`shell script.

The restore operation all the information contained in a backup which
includes:
 - database
 - files
 - plugins
 - themes
 - dotfiles

The operation will rename the existing files before restoring a backup,
which may tend to be useful in extreme cases. However, this should not
be considered as a substitute to recover a broken installation.

ALWAYS BACKUP THE CURRENT INSTALLATION USING the `app:backup:create`
CONTAINER ARGUMENT OR USING THE `redmine-backup-create` SCRIPT BEFORE
RESTORE ANOTHER BACKUP.

NOTES:
  - Backups can only be restored on the same or newer versions of
    Redmine. The inverse is not permitted.
  - The database adapter in use and the adapter used while creating the
    backup should be the same.